isolate and test the license file detection
Closes https://github.com/davglass/license-checker/issues/142
