+++ lib/index.js
@@ -12,6 +12,7 @@ var read = require('read-installed');
 var chalk = require('chalk');
 var treeify = require('treeify');
 var license = require('./license');
+var licenseFiles = require('./license-files');
 var debug = require('debug');
 var mkdirp = require('mkdirp');
 var spdxSatisfies = require('spdx-satisfies');
@@ -115,7 +116,7 @@ var flatten = function(options) {
     }
 
     licenseData = json.license || json.licenses || undefined;
-    
+
     if (json.path && (!json.readme || json.readme.toLowerCase().indexOf('no readme data found') > -1)) {
         readmeFile = path.join(json.path, 'README.md');
         /*istanbul ignore if*/
@@ -155,20 +156,7 @@ var flatten = function(options) {
     /*istanbul ignore else*/
     if (json.path && fs.existsSync(json.path)) {
         dirFiles = fs.readdirSync(json.path);
-
-        // Find and list license files in the precedence order
-        ['LICENSE', 'LICENCE', 'COPYING', 'README'].forEach(function(licenseFilename) {
-            var found = false;
-            dirFiles.forEach(function(filename) {
-                if (!found) {
-                    var basename = path.basename(filename, path.extname(filename)).toUpperCase();
-                    if (basename === licenseFilename) {
-                        files.push(filename);
-                        found = true;
-                    }
-                }
-            });
-        });
+        files = licenseFiles(dirFiles);
 
         noticeFiles = dirFiles.filter(function(filename) {
             filename = filename.toUpperCase();
